// src/app/shared/enums.ts
export enum StatusMessages {
  loading = 'Loading',
  needApiKey = 'Need API key',
  noDroplet = 'Ready to create',
  creating = 'Creating droplet',
  booting = 'Booting',
  vpnInstalling = 'Installing VPN',
  readyToConnect = 'Ready to connect',
  connecting = 'Connecting',
  connected = 'Connected',
  disconnecting = 'Disconnecting',
  destroying = 'Destroying'
}